rfc4180 (dsv->scm): Bugfix
The library would always fail to correctly parse RFC4180 format because there
was an error in the state machine: it considered even number of double quotes
as a proof that the input data was correct.

This change fixes the problem (#7 on GitHub) by re-writing the FSM to follow
the standard more closely. Also now the FSM for RFC4180 reads data character
by character which is more elegant and controllable.

Thanks to Alex Sassmannshausen for attracting my attention to the parser
issues in <https://github.com/artyom-poptsov/guile-dsv/issues/7>

* modules/dsv/rfc4180.scm (dsv->scm): Implement new finite state machine.
* tests/rfc4180.scm: Update tests.